+++ backend/Hubla.Sales.Application/Shared/Sales/UseCases/Outputs/SaleOutputBase.cs
@@ -9,10 +9,10 @@ namespace Hubla.Sales.Application.Shared.Sales.UseCases.Outputs
         public SaleType SaleType { get; set; }
         public DateTime Date { get; set; }
         public string Description { get; set; }
-        public double Value { get; set; }
+        public decimal Value { get; set; }
         public SellerOutputBase Seller { get; set; }
 
-        protected SaleOutputBase(int id, SaleType saleType, DateTime date, string description, double value, SellerOutputBase seller)
+        protected SaleOutputBase(int id, SaleType saleType, DateTime date, string description, decimal value, SellerOutputBase seller)
         {
             Id = id;
             SaleType = saleType;
@@ -21,7 +21,7 @@ namespace Hubla.Sales.Application.Shared.Sales.UseCases.Outputs
             Value = value;
             Seller = seller;
         }
-        public static SaleOutputBase Create(int id, SaleType saleType, DateTime date, string description, double value, SellerOutputBase seller) =>
+        public static SaleOutputBase Create(int id, SaleType saleType, DateTime date, string description, decimal value, SellerOutputBase seller) =>
             new(id, saleType, date, description, value, seller);
     }
 }
